generator/visual_creator.py
# Exemple minimal de génération visuelle
def create_visual(content, filename="output.png"):
    """
    Crée un fichier visuel basé sur le contenu fourni.
    """
    # Ici tu pourrais utiliser matplotlib ou PIL
    from PIL import Image, ImageDraw, ImageFont

    img = Image.new('RGB', (600, 200), color=(73, 109, 137))
    d = ImageDraw.Draw(img)
    d.text((10,10), content, fill=(255,255,0))
    img.save(filename)
    return filename