requirements.txt
absl-py==2.3.1
accelerate==1.8.0
alembic==1.17.1
altair==5.5.0
annotated-types==0.7.0
anyio==4.11.0
asttokens==3.0.0
async-timeout==5.0.1
attrs==25.4.0
blinker==1.9.0
brotli==1.2.0
cachetools==6.2.1
certifi==2025.11.12
charset-normalizer==3.4.4
click==8.3.0
cloudpickle==3.0.0
cmaes==0.12.0
colorlog==6.10.1
comm==0.2.3
contourpy==1.3.3
cycler==0.12.1
debugpy==1.8.17
decorator==5.2.1
executing==2.2.1
feather-format==0.4.1
filelock==3.20.0
fonttools==4.60.1
fsspec==2025.10.0
geomloss==0.2.6
gitdb==4.0.12
GitPython==3.1.45
greenlet==3.2.4
grpcio==1.76.0
h11==0.16.0
h5py==3.15.1
hf-xet==1.2.0
httpcore==1.0.9
httpx==0.28.1
huggingface_hub==1.1.4
idna==3.11
importlib-metadata==6.7.0
inflate64==1.0.3
ipykernel==7.1.0
ipython==9.7.0
ipython_pygments_lexers==1.1.1
jedi==0.19.2
Jinja2==3.1.6
joblib==1.5.2
jsonpatch==1.33
jsonpointer==3.0.0
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
jupyter_client==8.6.3
jupyter_core==5.9.1
kiwisolver==1.4.9
langchain==1.0.2
langchain-core==1.0.4
langchain-monorepo @ git+https://github.com/langchain-ai/langchain.git@ee19a30dde66674cbad2be0291ca805125823882
langgraph==1.0.3
langgraph-checkpoint==3.0.1
langgraph-prebuilt==1.0.2
langgraph-sdk==0.2.9
langsmith==0.4.42
llvmlite==0.45.1
loguru==0.7.2
Mako==1.3.10
Markdown==3.10
MarkupSafe==3.0.3
matplotlib==3.10.7
matplotlib-inline==0.2.1
mpmath==1.3.0
multivolumefile==0.2.3
narwhals==2.11.0
nest-asyncio==1.6.0
networkx==3.5
nflows==0.14
numba==0.62.1
numpy==2.3.4
optuna==3.1.0
orjson==3.11.4
ormsgpack==1.12.0
packaging==25.0
pandas==2.3.3
parso==0.8.5
pexpect==4.9.0
pillow==12.0.0
platformdirs==4.5.0
prompt_toolkit==3.0.52
protobuf==6.33.0
psutil==7.1.3
ptyprocess==0.7.0
pure_eval==0.2.3
py7zr==1.0.0
pyarrow==21.0.0
pybcj==1.0.6
pycox==0.2.3
pycryptodomex==3.23.0
pydantic==2.12.4
pydantic_core==2.41.5
pydeck==0.9.1
Pygments==2.19.2
pyparsing==3.2.5
pyppmd==1.2.0
python-dateutil==2.9.0.post0
pytz==2025.2
PyYAML==6.0.3
pyzmq==27.1.0
pyzstd==0.18.0
redis==5.0.1
referencing==0.37.0
requests==2.32.5
requests-toolbelt==1.0.0
rpds-py==0.28.0
safetensors==0.6.2
scikit-learn==1.7.2
scipy==1.16.3
shap==0.44.1
shellingham==1.5.4
six==1.17.0
slicer==0.0.7
smmap==5.0.2
sniffio==1.3.1
SQLAlchemy==2.0.44
stack-data==0.6.3
streamlit==1.51.0
sympy==1.14.0
synthcity==0.2.12
tenacity==8.3.0
tensorboard==2.20.0
tensorboard-data-server==0.7.2
texttable==1.7.0
threadpoolctl==3.6.0
toml==0.10.2
torch==2.1.0
torchtuples==0.2.2
torchvision==0.16.0
tornado==6.5.2
tqdm==4.67.1
traitlets==5.14.3
typer-slim==0.20.0
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.2
urllib3==2.5.0
watchdog==6.0.0
wcwidth==0.2.14
Werkzeug==3.1.3
xgboost==1.7.6
xxhash==3.6.0
zipp==3.23.0
zstandard==0.25.0
